Creates Rental model and updates video and customer models for relationships
