+++ app/__init__.py
@@ -32,10 +32,13 @@ def create_app(test_config=None):
     migrate.init_app(app, db)
 
     #Register Blueprints Here
-    from .routes import videos_bp
+    from .routes.video_routes import videos_bp
     app.register_blueprint(videos_bp)
 
-    from .routes_folder.customer_routes import customer_bp
+    from .routes.customer_routes import customer_bp
     app.register_blueprint(customer_bp)
+
+    from .routes.rental_routes import rental_bp
+    app.register_blueprint(rental_bp)
 
     return app
