+++ app/models/customer.py
@@ -6,7 +6,10 @@ class Customer(db.Model):
     postal_code = db.Column(db.String)
     phone = db.Column(db.String)
     registered_at = db.Column(db.DateTime, default=None, nullable=True)
-
+    # Many to many: list of videos that are rented. Only listed once
+    rented_videos = db.relationship('Video', secondary='rental')
+    rentals = db.relationship('Rental')
+    
     def to_dict(self):
         return {
             "id": self.id,
