+++ app/models/video.py
@@ -5,7 +5,11 @@ class Video(db.Model):
     title = db.Column(db.String)
     release_date = db.Column(db.DateTime)
     total_inventory = db.Column(db.Integer)
-
+    # Many to many: only counts each customer once. Secondary joins the third table
+    # backref - joins two tables
+    renters = db.relationship('Customer', secondary='rental')
+    # # One to many. Counts for all videos currently checked out
+    rentals = db.relationship('Rental')
 
     def to_dict(self):
         return {
