+++ app/routes/customer_routes.py
